requirements.txt
minio==7.2.15
pandas==2.2.3
python-dotenv==1.1.0
validators==0.35.0
